Move diagnose node to its own file
